+++ coeus-autograd/src/ops/nn/conv/mod.rs
@@ -9,5 +9,8 @@ mod utils;
 pub use conv1d::conv1d;
 pub use conv2d::conv2d;
 pub use conv3d::conv3d;
-pub use transpose::{conv_transpose1d, conv_transpose2d, ConvTranspose1dNode, ConvTranspose2dNode};
+pub use transpose::{
+    conv_transpose1d, conv_transpose2d, conv_transpose3d, ConvTranspose1dNode, ConvTranspose2dNode,
+    ConvTranspose3dNode,
+};
 pub use utils::ConvNode;