+++ coeus-autograd/src/ops/nn/mod.rs
@@ -18,7 +18,7 @@ pub mod pool;
 pub mod softmax;
 
 pub use attention::{sdp_attention, AttentionMask, CausalMask, NullMask};
-pub use conv::{conv1d, conv2d, conv3d, conv_transpose1d, conv_transpose2d};
+pub use conv::{conv1d, conv2d, conv3d, conv_transpose1d, conv_transpose2d, conv_transpose3d};
 pub use dropout::dropout;
 pub use log_softmax::log_softmax;
 pub use loss::{